+++ lib/parser/errors.go
@@ -25,7 +25,7 @@ func (e EmError) Error() string {
 	formatPos := func(pos Position) string {
 		//fmt.Printf("%d %d\n", pos.Offset, pos.Line)
 		if pos.Offset > 0 && pos.Line > -1 {
-			return fmt.Sprintf("line #%d col #%d", e.Line+e.Offset, e.Column)
+			return fmt.Sprintf("line %d col %d", e.Line+e.Offset, e.Column)
 		}
 		return ""
 	}